+++ .clang-format
@@ -5,7 +5,7 @@ IncludeCategories:
     Priority:        4
   - Regex:           '^"(llvm|llvm-c|clang|clang-c|mlir|mlir-c)/'
     Priority:        3
-  - Regex:           '^"(cudaq|\.\.)/'
+  - Regex:           '^"(qoda|\.\.)/'
     Priority:        2
   - Regex:           '.*'
     Priority:        1
