--- python/tests/compiler/qalloc_empty.py
@@ -1,40 +1,0 @@
-# ============================================================================ #
-# Copyright (c) 2022 - 2023 NVIDIA Corporation & Affiliates.                   #
-# All rights reserved.                                                         #
-#                                                                              #
-# This source code and the accompanying materials are made available under     #
-# the terms of the Apache License 2.0 which accompanies this distribution.     #
-# ============================================================================ #
-
-# RUN: PYTHONPATH=../../ pytest -rP  %s | FileCheck %s
-
-import os
-
-import pytest
-import numpy as np
-
-import cudaq
-
-
-
-def test_kernel_qalloc_empty():
-    """
-    Test `cudaq.Kernel.qalloc()` when no arguments are provided.
-    """
-    kernel = cudaq.make_kernel()
-    # Use `qalloc()` with no function arguments.
-    qubit = kernel.qalloc()
-    # Assert that only 1 qubit is allocated in the MLIR.
-    print(kernel)
-
-
-# CHECK-LABEL:   func.func @__nvqpp__mlirgen____nvqppBuilderKernel_{{.*}}() {
-# CHECK:           %[[VAL_0:.*]] = quake.alloca : !quake.qref
-# CHECK:           return
-# CHECK:         }
-
-
-# leave for gdb debugging
-if __name__ == "__main__":
-    loc = os.path.abspath(__file__)
-    pytest.main([loc, "-rP"])
