+++ runtime/cudaq/platform/quantum_platform.cpp
@@ -16,10 +16,10 @@
 #include "cudaq/platform/quantum_platform.h"
 #include "common/Logger.h"
 #include "common/PluginUtils.h"
-#include "nvqpp_config.h"
 #include "cudaq/platform/qpu.h"
 #include "cudaq/qis/qubit_qis.h"
 #include "cudaq/qis/qudit.h"
+#include "nvqpp_config.h"
 #include <fmt/core.h>
 #include <fstream>
 #include <iostream>
