+++ tools/qpud/qpud.cpp
@@ -10,12 +10,12 @@
 #include "NvidiaPlatformHelper.h"
 #include "TargetBackend.h"
 #include "common/Logger.h"
-#include "nvqpp_config.h"
-#include "rpc/server.h"
-#include "rpc/this_handler.h"
 #include "cudaq/Optimizer/Dialect/CC/CCDialect.h"
 #include "cudaq/Optimizer/Dialect/Quake/QuakeDialect.h"
 #include "cudaq/utils/cudaq_utils.h"
+#include "nvqpp_config.h"
+#include "rpc/server.h"
+#include "rpc/this_handler.h"
 #include "llvm/Support/TargetSelect.h"
 #include "mlir/Dialect/Affine/IR/AffineOps.h"
 #include "mlir/Dialect/Arith/IR/Arith.h"
